requirements.txt
django
djangorestframework
pillow
psycopg2-binary
django-decouple
djangorestframework-simplejwt
celery